+++ backend/src/gitinfo.ts
@@ -1,7 +1,7 @@
 export const GitInfo = {
-  commitHash: "6558ef8",
-  commitTimestamp: "2026-07-19T08:43:55-03:00",
+  commitHash: "f6f6d37",
+  commitTimestamp: "2026-07-19T11:16:54-03:00",
   branchName: "main",
   tagName: "v1.0.x",
-  buildTimestamp: "2026-07-19T11:43:59Z"
+  buildTimestamp: "2026-07-19T14:17:46Z"
 };